migrated util code where applicable
