+++ lib/utils/data_utils.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var moment = require('moment');
-var dateFormat = require('./elastic_utils').dateFormat;
+var dateFormat = require('./date_utils').dateFormat;
 
 
 module.exports = function(opConfig, otherSchema) {
